requirements.txt
pillow==10.1.0
numpy==1.26.0
pandas==2.1.2
requests==2.31.0
python-dotenv==1.0.0
tqdm==4.66.1
matplotlib==3.8.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
EOL
ImageHash==4.3.1
opencv-python==4.8.1.78